requirements.txt
# Core dependencies
mcp[cli]>=1.3.0    # MCP核心功能和CLI工具
telnetlib3>=2.0.0  # 用于telnet连接
psutil>=5.9.0      # 用于进程管理

# Development dependencies
pytest>=7.0.0      # 测试框架
black>=23.0.0      # 代码格式化
isort>=5.0.0       # import排序
mypy>=1.0.0        # 类型检查
